+++ routes/web.php
@@ -18,25 +18,23 @@ use Illuminate\Support\Facades\Route;
 Route::get('download/file/{roomFile}/{filename?}', [FileController::class,'show'])->name('download.file')->middleware('signed');
 Route::get('download/attendance/{meeting}', [MeetingController::class,'attendance'])->name('download.attendance')->middleware('auth:users,ldap');
 
-Route::middleware('csp')->group(function () {
-    if (config('greenlight.compatibility')) {
-        Route::prefix(config('greenlight.base'))->group(function () {
-            // room urls
-            Route::get('/{id}', function ($id) {
-                return redirect('/rooms/'.$id);
-            })->where('id', '([A-Za-z0-9]{3}-[A-Za-z0-9]{3}-[A-Za-z0-9]{3}(-[A-Za-z0-9]{3})?)');
-            // login
-            Route::redirect('/ldap_signin', '/login');
-            Route::redirect('/signin', '/login');
-            // default room
-            Route::redirect('/default_room', '/rooms/own');
-            // all other routes
-            Route::redirect('/', '/');
-            Route::redirect('/{any}', '/')->where('any', '.*');
-        });
-    }
-
-    if (!env('DISABLE_CATCHALL_ROUTES')) {
-        Route::any('/{any}', [ApplicationController::class,'index'])->where('any', '.*');
+if (config('greenlight.compatibility')) {
+    Route::prefix(config('greenlight.base'))->group(function () {
+        // room urls
+        Route::get('/{id}', function ($id) {
+            return redirect('/rooms/'.$id);
+        })->where('id', '([A-Za-z0-9]{3}-[A-Za-z0-9]{3}-[A-Za-z0-9]{3}(-[A-Za-z0-9]{3})?)');
+        // login
+        Route::redirect('/ldap_signin', '/login');
+        Route::redirect('/signin', '/login');
+        // default room
+        Route::redirect('/default_room', '/rooms/own');
+        // all other routes
+        Route::redirect('/', '/');
+        Route::redirect('/{any}', '/')->where('any', '.*');
+    });
 }
-});
+
+if (!env('DISABLE_CATCHALL_ROUTES')) {
+    Route::any('/{any}', [ApplicationController::class,'index'])->where('any', '.*');
+}